Revert f32 -> f64

+++ src/c2.rs
@@ -5,26 +5,26 @@ use na::{Complex, Vector2, Vector4, ComplexField, Matrix2};
 /// A complex 2-vector, i.e., a member of the vector space C^2.
 #[derive(Clone, Copy, Debug)]
 pub struct C2 {
-    pub vec: Vector2<Complex<f32>>
+    pub vec: Vector2<Complex<f64>>
 }
 
 impl C2 {
     /// Create the C^2 vector (x, y) from x and y.
-    pub fn new(x: Complex<f32>, y: Complex<f32>) -> Self {
+    pub fn new(x: Complex<f64>, y: Complex<f64>) -> Self {
         Self {
             vec: Vector2::new(x, y)
         }
     }
 
     /// Create the C^2 vector (a + bi, c + di) for real a, b, c, d.
-    pub fn from_components(a: f32, b: f32, c: f32, d: f32) -> Self {
+    pub fn from_components(a: f64, b: f64, c: f64, d: f64) -> Self {
         Self {
             vec: Vector2::new(Complex::new(a, b), Complex::new(c, d))
         }
     }
 
     /// Convert (a + bi, c + di) to the 4-tuple of floats (a, b, c, d).
-    pub fn to_components(&self) -> (f32, f32, f32, f32) {
+    pub fn to_components(&self) -> (f64, f64, f64, f64) {
         (
             self.vec.x.re,
             self.vec.x.im,
@@ -34,7 +34,7 @@ impl C2 {
     }
 
     /// Create the C^2 vector (a + bi, c + di) from the R^4 vector (a, b, c, d).
-    pub fn from_vector4(vec: &Vector4<f32>) -> Self {
+    pub fn from_vector4(vec: &Vector4<f64>) -> Self {
         Self {
             vec: Vector2::new(
                 Complex::new(vec.x, vec.y),
@@ -48,7 +48,7 @@ impl C2 {
     }
 
     /// Convert the C^2 vector (a + bi, c + di) to the R^4 vector (a, b, c, d).
-    pub fn to_vector4(&self) -> Vector4<f32> {
+    pub fn to_vector4(&self) -> Vector4<f64> {
         Vector4::new(
             self.vec.x.re,
             self.vec.x.im,
@@ -58,28 +58,28 @@ impl C2 {
     }
 
     /// Return the norm: ||(x, y)|| = sqrt(|x|^2 + |y|^2).
-    pub fn abs(&self) -> f32 {
+    pub fn abs(&self) -> f64 {
         self.vec.norm()
     }
 
     /// Return the absolute difference between two C2 vectors.
-    pub fn abs_difference(&self, other: &Self) -> f32 {
+    pub fn abs_difference(&self, other: &Self) -> f64 {
         (self.vec - other.vec).norm()
     }
 
     /// Take the inner product <(x1, x2), (y1, y2)> = x1 conj(y1) + x2 conj(y2), which is linear in
     /// the first argument.
-    pub fn inner(&self, other: &Self) -> Complex<f32> {
+    pub fn inner(&self, other: &Self) -> Complex<f64> {
         other.vec.dotc(&self.vec)
     }
 
     /// Absolute value of inner product squared.
-    pub fn inner_abs_squared(&self, other: &Self) -> f32 {
+    pub fn inner_abs_squared(&self, other: &Self) -> f64 {
         self.inner(&other).modulus_squared()
     }
 
     /// Absolute value of inner product.
-    pub fn inner_abs(&self, other: &Self) -> f32 {
+    pub fn inner_abs(&self, other: &Self) -> f64 {
         self.inner(&other).abs()
     }
 
@@ -92,7 +92,7 @@ impl C2 {
 
     /// Given this vector x, return a special 2x2 unitary matrix M in SU(2) such that Mx = (k, 0)
     /// for real k.
-    pub fn normalizing_su2_matrix(&self) -> Matrix2<Complex<f32>> {
+    pub fn normalizing_su2_matrix(&self) -> Matrix2<Complex<f64>> {
         let norm = self.abs();
         Matrix2::new(
             self.vec.x.conj() / norm,
@@ -103,34 +103,34 @@ impl C2 {
     }
 
     /// Return the inverse of self.normalizing_su2_matrix.
-    pub fn normalizing_su2_matrix_inv(&self) -> Matrix2<Complex<f32>> {
+    pub fn normalizing_su2_matrix_inv(&self) -> Matrix2<Complex<f64>> {
         self.normalizing_su2_matrix().adjoint()
     }
 
     /// Given C^2 vectors z1 and z2, return the similarity |<z1, z2>| / (||z1|| ||z2||). This is a
     /// cosine-like similarity function which is 1 iff they are phase rotations of each other.
-    pub fn similarity(&self, other: &Self) -> f32 {
+    pub fn similarity(&self, other: &Self) -> f64 {
         self.inner_abs(&other) / (self.abs() * other.abs())
     }
 }
 
 
-impl Mul<f32> for &C2 {
+impl Mul<f64> for &C2 {
     type Output = C2;
 
     /// Multiply by a real scalar.
-    fn mul(self, rhs: f32) -> C2 {
+    fn mul(self, rhs: f64) -> C2 {
         C2 {
             vec: self.vec * Complex::from_real(rhs)
         }
     }
 }
 
-impl Div<f32> for &C2 {
+impl Div<f64> for &C2 {
     type Output = C2;
 
     /// Divide by a real scalar.
-    fn div(self, rhs: f32) -> C2 {
+    fn div(self, rhs: f64) -> C2 {
         C2 {
             vec: self.vec / Complex::from_real(rhs)
         }
@@ -166,7 +166,7 @@ mod test {
     #[test]
     fn test_from_to_vector4() {
         let (a, b, c, d) = (1.0, 2.0, -3.0, 4.0);
-        let vector4 = Vector4::<f32>::new(a, b, c, d);
+        let vector4 = Vector4::<f64>::new(a, b, c, d);
         let x = C2::from_vector4(&vector4);
         assert_eq!(
             x.vec,
